+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.2.26
 - Fix eggs not being picked up by container auto-pickup [#67](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/67)
+- Fix general issues with containers and auto-pickup
 
 ### v0.2.25
 - New feature: